+++ src/main/java/Server/PacketBroker.java
@@ -1,0 +1,42 @@
+package Server;
+
+import Util.Network.Packet;
+import Util.SocketProxy;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.concurrent.BlockingQueue;
+
+/// Verteilt Pakete an alle angemeldeten Clients
+public class PacketBroker implements Runnable {
+	private final BlockingQueue<Packet> packetBrokerQueue;
+	private final List<SocketProxy> clients;
+
+	public PacketBroker(BlockingQueue<Packet> packetBrokerQueue, List<SocketProxy> clients) {
+		this.packetBrokerQueue = packetBrokerQueue;
+		this.clients = clients;
+	}
+
+	@Override
+	public void run() {
+		while (true) {
+			try {
+				Packet packet = packetBrokerQueue.take();
+
+				synchronized (clients) {
+					for (SocketProxy client : clients) {
+						try {
+							client.out.writeObject(packet);
+							client.out.flush();
+						} catch (IOException e) {
+							System.out.println("Fehler beim Senden:\n" + e);
+						}
+					}
+				}
+
+			} catch (InterruptedException e) {
+				break;
+			}
+		}
+	}
+}
