+++ src/main/java/Server/Server.java
@@ -1,7 +1,9 @@
 package Server;
 
-import Util.Network.Messages.Message;
+import Util.Network.Notifications.JoinNotification;
+import Util.Network.Packet;
 import Util.SocketProxy;
+import Util.User;
 
 import java.io.IOException;
 import java.net.ServerSocket;
@@ -14,19 +16,18 @@ import java.util.concurrent.BlockingQueue;
 
 public class Server implements Runnable {
 	private final ServerSocket server;
-
-	private final BlockingQueue<Message> messageBrokerQueue;
+	private final BlockingQueue<Packet> packetBrokerQueue;
 	private final List<SocketProxy> clients;
 
 	public Server(int port) throws IOException {
 		server = new ServerSocket(port);
-		messageBrokerQueue = new ArrayBlockingQueue<>(32);
+		packetBrokerQueue = new ArrayBlockingQueue<>(32);
 		clients = Collections.synchronizedList(new ArrayList<>());
 	}
 
 	@Override
 	public void run() {
-		new Thread(new MessageBroker(messageBrokerQueue, clients)).start();
+		new Thread(new PacketBroker(packetBrokerQueue, clients)).start();
 
 		while (true) {
 			try {
@@ -36,11 +37,16 @@ public class Server implements Runnable {
 
 				clients.add(client);
 
-				new Thread(new ClientHandler(client, messageBrokerQueue)).start();
+				// todo: Sobald es ein Loginsystem gibt, hier den Benutzernamen des verbindenden Clients übergeben
+				packetBrokerQueue.put(new JoinNotification(new User("Platzhalter")));
+
+				new Thread(new ClientHandler(client, packetBrokerQueue)).start();
 
 			} catch (IOException e) {
 				System.out.println("Fehler beim Verbindungsaufbau:\n" + e);
 				break;
+			} catch (InterruptedException e) {
+				throw new RuntimeException(e);
 			}
 		}
 	}
