+++ src/main/java/Util/Network/Messages/EmojiMessage.java
@@ -1,10 +1,8 @@
-package Util.Emoji;
+package Util.Network.Messages;
 
-import Util.Network.Messages.Message;
 import Util.User;
 
 public class EmojiMessage extends Message {
-
     private final String emoji;
 
     public EmojiMessage(User sender, String emoji) {
